third_party/blink/web_tests/platform/generic/fast/forms/month-multiple-fields/month-multiple-fields-preserve-value-after-history-back-expected.txt

Checks input value is preserved after backward and forward

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Went back to a page. Checking a form control which had a full value:
PASS "12012-10" is "12012-10"
Went back to a page. Checking a form control which had a partial value:
PASS "January ----" is "January ----"
Went forward to a page. Checking a form control which had a full value:
PASS "0001-11" is "0001-11"
Went forward to a page. Checking a form control which had a partial value:
PASS "--------- 1999" is "--------- 1999"


PASS successfullyParsed is true

TEST COMPLETE
